Backend/rotary/requirements.txt
pymongo==4.6.2
spidev==3.6
RPi.GPIO==0.7.1